requirements.txt
fastapi>=0.110
uvicorn>=0.29
httpx>=0.26
pydantic>=2.7
fastapi
uvicorn
python-dotenv
pymongo
requests